+++ apps/chat-e2e/src/ui/webElements/promptBar.ts
@@ -1,5 +1,6 @@
-import { SideBarSelectors } from '../selectors';
+import { PromptBarSelectors, SideBarSelectors } from '../selectors';
 
+import { Styles, removeAlpha } from '@/src/ui/domData';
 import {
   FolderPrompts,
   PromptsTree,
@@ -56,9 +57,29 @@ export class PromptBar extends SideBar {
     return this.sharedFolderPrompts;
   }
 
+  public newEntityButton = this.getChildElementBySelector(
+    PromptBarSelectors.newEntity,
+  );
+
   public async createNewPrompt() {
     await this.newEntityButton.waitForState();
     await this.newEntityButton.click();
+  }
+
+  public async hoverOverNewEntity() {
+    await this.newEntityButton.waitForState();
+    await this.newEntityButton.hoverOver();
+  }
+
+  public async getNewEntityBackgroundColor() {
+    const backgroundColor = await this.newEntityButton.getComputedStyleProperty(
+      Styles.backgroundColor,
+    );
+    return removeAlpha(backgroundColor[0]);
+  }
+
+  public async getNewEntityCursor() {
+    return this.newEntityButton.getComputedStyleProperty(Styles.cursor);
   }
 
   public async dragAndDropPromptFromFolder(
